CompLast {RHNetTools}	R Documentation

Returns a list of M,matrices with the same dimensions as the input 'n'. For all row/col indices that are identified in 'x' as missing for a given time point, the entire row/col matching the index is set to the last known value.

Description

Returns a list of M,matrices with the same dimensions as the input 'n'. For all row/col indices that are identified in 'x' as missing for a given time point, the entire row/col matching the index is set to the last known value.

Usage


CompLast(x, n)

Arguments

x
A data frame with k + 1 columns, where k == length(n), and the first column is an identifier column. Columns 2:ncol(x) must correspond with matrices 1:length(n).

n
A list of matrices (representing networks across time). Matrices should be identical in dimensionality and should include row and column names that match with the identifiers in the input x.

Value

A list of n matrices with the exact structure of the input 'n', in which rows corresponding with missing participants, as captured in the input 'x', set to all NA values.

[Package RHNetTools version 2.1.3 Index]