CompMod {RHNetTools}	R Documentation

Returns a list of M,matrices with the same dimensions as the input 'n' except with entire rows set to NA for those rows with names matching the names of rows in the input 'x', which identifies time points when an actor was missing (NA) from the network data.

Description

Returns a list of M,matrices with the same dimensions as the input 'n' except with entire rows set to NA for those rows with names matching the names of rows in the input 'x', which identifies time points when an actor was missing (NA) from the network data.

Usage


CompMod(n, x, test = NA, value = NA)

Arguments

n
A list of matrices (representing networks across time). Matrices should be identical in dimensionality and should include row and column names that match with the identifiers in the input x.

x
A data frame with k + 1 columns, where k == length(n), and the first column is an identifier column. Columns 2:ncol(x) must correspond with matrices 1:length(n). This is usually "SIDInAnyWave"

value
The value to plug in (e.g., NA, or 10)

Value

A list of n matrices with the exact structure of the input 'n', in which rows corresponding with missing participants, as captured in the input 'x', set to all NA values.

[Package RHNetTools version 2.1.3 Index]