DeltaNet {RHNetTools}	R Documentation

Given a list of networks representing time ordered observations of a network, calculate aggregate change statistics from one observation to the next.

Description

Given a list of networks representing time ordered observations of a network, calculate aggregate change statistics from one observation to the next.

Usage


DeltaNet(n)

Arguments

n
A list of networks as, e.g., output by makeNetworkSet()

Value

A table characterizing network changes.

[Package RHNetTools version 2.1.3 Index]