QOLMixing {RHNetTools}	R Documentation

RSiena goodness-of-fit statistic function for a 3x3 mixing matrix of QOL Few details are supplied because this function is not intended to be used directly. Moreover, this is a specialized function for a specific project and a specific covariate.

Description

RSiena goodness-of-fit statistic function for a 3x3 mixing matrix of QOL Few details are supplied because this function is not intended to be used directly. Moreover, this is a specialized function for a specific project and a specific covariate.

Usage


QOLMixing(i, data, sims, wave, groupName, varName, levls = 1:9)

[Package RHNetTools version 2.1.3 Index]