TriadCensus {RHNetTools}	R Documentation

RSiena goodness-of-fit statistic function for the 16-dimensional triad census. Few details are supplied because this function is not intended to be used directly.

Description

RSiena goodness-of-fit statistic function for the 16-dimensional triad census. Few details are supplied because this function is not intended to be used directly.

Usage


TriadCensus(i, data, sims, wave, groupName, varName, levls = 1:16)

[Package RHNetTools version 2.1.3 Index]