addStrZeros.1 {RHNetTools}	R Documentation

Insert structural zeros to prevent ties between houses when two nodes have never lived in the same house at the same time.

Description

Insert structural zeros to prevent ties between houses when two nodes have never lived in the same house at the same time.

Usage


addStrZeros.1(pNetList, pSIDHIDList)

Arguments

pNetList
A list of social network matrices (typically for a set of waves). * Class may be + 'matrix' (base R 2-dim array) + 'Matrix' (any of the s4 classes supported by pkg 'Matrix') * Rows and columns must be named by node identifier (usually SID)

pSIDHIDList
A data.frame with ncol() == the number of waves + 1. The first col is the participant ID, and the remaining cols are house identifiers representing where the participant lived at each observation.

Value

Returns the list of network matrices in the same form they were input, but with structural zeros (10's) encoded for linkages between individuals in different houses.

[Package RHNetTools version 2.1.3 Index]