addStrZeros.2 {RHNetTools}	R Documentation

Insert structural zeros to prevent ties between houses.

Description

Insert structural zeros to prevent ties between houses.

Usage


addStrZeros.2(pNetList, pSIDHIDList)

Arguments

pNetList
A list of social network matrices (typically for a set of waves). * Class may be + 'matrix' (base R 2-dim array) + 'Matrix' (any of the s4 classes supported by pkg 'Matrix') * Rows and columns must be named by node identifier (usually SID)

pSIDHIDList
A list of dataframes (or objects that can act like them, such as data.tables), corresponding to the networks in pNetList. Each is a lookup table that codes which nodes (SIDs) belong in which houses (HIDs) so each has two columns (SID and HID), which index this correspondence.

Value

Returns the list of network matrices in the same form they were input, but with structural zeros (10's) encoded for linkages between individuals in different houses.

[Package RHNetTools version 2.1.3 Index]