checkForHousesOf1 {RHNetTools}	R Documentation

Check for houses with just 1 network member

Description

Check for houses with just 1 network member

Usage


checkForHousesOf1(rlpNet)

Arguments

rlpNet
a tbl_df in 'longNet' or 'rlpNet format. This table can contain any number of waves, but if there is more than one wave, the check will be done within wave, i.e. so the user can be sure that no single wave has a house with just one member surveyed.

Details

In such cases there is no 'network', so they should be excluded from any network analysis. Usually this should not arise but it can if individuals were excluded from a supposed-final analysis set because, e.g., they did not have enough observations (say, 2 or more across all waves) to be included.

Value

A tbl_df containing WID, and HID for any individuals & waves where the individual was the only one in his/her house.

Note

This check is important to run before taking any network set into production analyses, because other data corrections can create houses of size 1 as a 'side effect' that may not otherwise be obvious.

[Package RHNetTools version 2.1.3 Index]