checkForNoAlters {RHNetTools}	R Documentation

Check individuals who were only egos in a house (never alters).

Description

Check individuals who were only egos in a house (never alters).

Usage


checkForNoAlters(rlpNet, pWave = 0)

Arguments

rlpNet
(tbl_df) A tbl in 'longNet' or 'rlpNet' format. Can include multiple waves but only one will be processed (see 'pWave' doc).

pWave
(scaler, default=0) The wave to select. If default, function assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise the function will throw an error.

Details

Because house residents are asked to rate all other residents on study relationship criteria, there should never be a case where someone did a survey, but was not named as an alter by anyone. This check also covers the situation where only 1 individual was surveyed in a house (in that instance, there will be just 1 ego, though possibly multiple alters)

Value

A tbl_df containing SID, HID, and WID for any individuals & waves where the individual did a survey, but never appears as an alter (AltID) in the input table.

Note

Only within a wave does it make sense to assume that all egos should be chosen by at least some (same-house) alters.

[Package RHNetTools version 2.1.3 Index]