checkForSelfSelection {RHNetTools}	R Documentation

Checks long-format data (from 'makeLongNet') for self-choices (loops)

Description

Self-choices are not allowed. RSiena ignores them, but they make graphs weird, and make the addition of structural zeros more difficult.

Usage


checkForSelfSelection(rlpNet, pWave = 0)

Arguments

rlpNet
(tbl_df)Input long-format, i.e. 'longNet' or 'rlpNet' format.

pWave
(scaler, default=0) The wave to select. If default, function assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise the function will throw an error.

Value

Outputs a tbl showing self-choices. Normally these should be deleted.

[Package RHNetTools version 2.1.3 Index]