datToSiena {RHNetTools}	R Documentation

Generate a "siena" data object from a set of properly formed data frames ("cx" and "vx") and a list of adjacency matrices ordered by wave.

Description

Generate a "siena" data object from a set of properly formed data frames ("cx" and "vx") and a list of adjacency matrices ordered by wave.

Usage


datToSiena(
  cx = NULL,
  vx = NULL,
  net = NULL,
  pidf = "SID",
  pid = NULL,
  depvar = NULL,
  concov = NULL,
  varcov = NULL,
  RSiena = T,
  nodeSet = "Actors",
  hid = NULL,
  vDyadCov = NULL,
  compChange = NULL,
  allowOnlyBeh = T,
  allowOnlyNet = T
)

Arguments

cx
Data frame of constant covariates with id field == pidf.

vx
List of data frames, one for each wave, containing identically named columns representing time-specific variables (including dependent behavioral variables).

net
Named list of lists of sparse matrices of type "dgTMatrix". The names given to each list are important and will be used as reference. If the list is unnamed, it is assumed to be a list of waves of one network.

pidf
Field name common to cx and vx data frames with participant identifiers. Default is "SID"

pid
Vector of IDs to retain. If NA, then retain all.

depvar
String vector of time varying variables that will be used as dependent behavioral variables.

concov
String vector of constant covariate field names to retain. If NA, then retain all.

varcov
Vector of time varying covariate field names to retain. If NA, then retain all.

RSiena
Should the output be formed into a data object that RSiena can work with? Otherwise, subsets are applied and data frames are returned

cextra
Fields in cx to be removed from the constant covariates, but still returned for later use (e.g., house id)

Details

An ideal use of this function is in the context of RSiena data preparation. At the least, the user supplies a data frame ("cx") containing all constant actor covariates; a list of data frames ("vx") containing wave-specific time-varying actor covariates–one data frame per wave with identical column names; a string ("pidf") referring to the name of the participant id field common to all data frames (normally "SID". Optionally, the user may supply a vector of participant ids ("pid") for subsetting all output data; a string vector of field names ("concov" and "varcov") that identify fields of interest from the input data frames (NA implies all are of interest); and a logical ("RSiena") indicating whether the result should be returned as a "siena" data object.

Value

If RSiena==TRUE, a list with three elements:
RSienaObj: a list of 'coCovar' and 'varCovar' objects amenable to coercion to a 'siena' object.
pid: a vector of pid values that were used to subset and order the output data
extra: a data frame of variables identified with the 'cextra' parameter that were not included in the RSienaObj, but may yet be needed subsetted and ordered as objects in the RSienaObj

Author(s)

Nathan J. Doogan doogan.1@osu.edu

[Package RHNetTools version 2.1.3 Index]