errChecks {RHNetTools}	R Documentation

Internal function to check common errors for these functions

Description

Internal function to check common errors for these functions

Usage


errChecks(rlpNet, pWave)

Arguments

rlpNet
(tbl_df) A network edge list in the form of rlpNet, or longNet. May be multiwave, but only 1 wave is ever used.

pWave
(scaler, default=0) The wave to select. If default, function assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise the function will throw an error.

[Package RHNetTools version 2.1.3 Index]