getNet4OneRlp {RHNetTools}	R Documentation

Returns an edgelist data.table for just 1 type of relationship

Description

Returns an edgelist data.table for just 1 type of relationship

Usage


getNet4OneRlp(tbLong, pWavVec = c(1), pRlpName)

Arguments

tbLong
A long-format tibble, output by 'makeLongNet'

pWavVec
A vector of waves; default is just the 1st wave

pRlpName
A character string with the name of the relationship; this must be a name in dtLong corresponing to a type of relationship. These are: friend, loan, help, conv, advc, rate.

Value

Returns an edgelist-format tibble, with the columns SID, HID, WID, and AltID for the selected relationship type. Will have the same number of rows as tbLong, less any for which the relationship selected for has NAs.

[Package RHNetTools version 2.1.3 Index]