makeAdjMatrix {RHNetTools}	R Documentation

Makes an adjacency matrix from an edge list, with SID-labeled rows/cols

Description

Makes an adjacency matrix from an edge list, with SID-labeled rows/cols

Usage


makeAdjMatrix(
  rlpNetAllWaves,
  pWave = 0,
  includeAltID = FALSE,
  not_active_code = 10
)

Arguments

rlpNetAllWaves
An edge list network of the form created by 'getNet4OneRlp', i.e. a tbl_df, with cols: SID, HID, WID, AltID, and one of: friend, loan, help, conv, advc, rate. If for one wave of a multiwave set, this tbl MUST include all edges for all waves (see 'details' for explanation).

pWave
A numeric value indicating which WID to use from rlpNet, if the latter includes multiple waves (which it should, generally, when creating network matrices for a multiwave analysis; see 'details' for more on this)

includeAltID
Default=FALSE. If TRUE, both AltIDs and SIDs are included in the output matrix, so that individuals who do not appear as a chooser (SID) might still be included in the output. If FALSE, only choosers (with an SID in rlpNet) will be included; any links to AltIDs who are not also SIDs are ignored.

not_active_code
Default=10. This is the code put into the adjacency matrix if the individual represented by that row & column did not contribute network data during wave 'pWave'

Details

The function does not break up rlpNet in any way; that is, it returns a single matrix, with 1's where SID chose AltID, and 0's otherwise.
*Importantly*, if this function is being used to create a network matrix for a particular wave of a multiwave analysis, and if also included SIDs do not necessarily appear in all waves (which is typical), the input tbl (rlpNetAllWaves) must include the data for the entire analysis, i.e. all edges for all waves.

Value

A matrix of dim unique(SID) (or if includeAltID=TRUE, of unique(SID U AltID)), containing the values appearing in col 5 of rlpNet, and with rownames = colnames = the original ID values (SID &/or AltID), to identify which rows and columns refer to which individuals in rlpNet.

Note

The function *does not* add structural zeros between individuals in different houses. For that, create a network set and input it to the function 's0Fill'; you get output in the same format, but with the structural zeros added.

[Package RHNetTools version 2.1.3 Index]