makeCCVec {RHNetTools}	R Documentation

Returns a list of vectors representing RSiena composition change.

Description

The output of this function may be input directly to the RSiena function 'sienaCompositionChange'.

Usage


makeCCVec(pElig)

Arguments

pElig
An n x 1+w tbl_df, with 1 row for every SID to be included in a particular analysis (an "Analysis Set"), and 1 col for each wave w, containing a number >0 if the SID is in the analysis for that wave, and NA if not if not (col 1 is SID). In this project, this dataframe (or data.table) is called 'SIDInAnyWave', or a similar construction, with SID as the first column, and the house id the SID belonged to in each of the w waves in the analysis, named 'HID1', 'HID2', ....'HIDw'.

Value

A list of n elements, one for each row of the input DF. Each element is a vector of pairs of arrival and departure times from the network, as explained in the RSiena manual under 'method of changing composition'.

Examples


# Returns a list of composition change vectors for the SIDs
# in waves 1, 3, and 4, and schools 3, 4, 5, 6, and 30.
ccDF <- makeCCVec(SIDInAnyWave)


[Package RHNetTools version 2.1.3 Index]