makeCountByHouseByWave {RHNetTools}	R Documentation

Make table showing number of residents x house x wave

Description

Make table showing number of residents x house x wave

Usage


makeCountByHouseByWave(longTB, includeAltID = FALSE, print2Cons = FALSE)

Arguments

longTB
(tbl_df) A multiwave long-format (from 'makeLongNet') tibble.

includeAltID
(logical) If TRUE, include individuals who appear as only altIDs in the counts (i.e. individuals who were residents at the time of survey, but who did not participate in the survey)

print2Cons
(logical) If TRUE, write the output to the console.

Details

Should be a method for the longTB.raw class, when there is one.

Value

(tbl_df) A tibble with cols HID, WID, Count, in order by HID, WID.

[Package RHNetTools version 2.1.3 Index]