makeListOfSIDsByHouse {RHNetTools}	R Documentation

Returns a list of labeled house elements, each containing a vector of the SIDs in that house for the wave specified.

Description

Returns a list of labeled house elements, each containing a vector of the SIDs in that house for the wave specified.

Usage


makeListOfSIDsByHouse(rlpNet, pWID = 1, includeAltID = FALSE)

Arguments

rlpNet
(tbl_df) Requires the format produced by 'getNet4OneRlp', i.e. with columns: SID, HID, WID, AltID, <rlpname>, where <rlpname> is one of: friend, loan, help, conv, advc, or rate. It would also work with any long-format edgelist that has the columns WID, SID, HID, and AltID.

pWID
(scaler, default 1) The wave number being processed

includeAltID
(logical, default FALSE)

Value

A list of vectors, each element containing the SIDs of the residents of the corresponding house. Elements are named by house number, and are ordered by house number.

[Package RHNetTools version 2.1.3 Index]