makeSIDInAnyWave {RHNetTools}	R Documentation

Creates a tibble of SIDs x houses in each wave

Description

Creates a tibble of SIDs x houses in each wave

Usage


makeSIDInAnyWave(pInLst)

Arguments

pInLst
(list) A list of tbls, one for each wave of interest, with columns SID and HID, e.g. as created by 'makeSIDHIDLookups'.

Details

If the individual was not a resident anywhere (i.e. not in the study), the value for that SID and wave is NA. If the individual switched houses while surveys were in progress, it is possible he/she could have had a longNet row from two houses; however, he/she could only have surveyed in one, and this is the one that should be assigned to that SID. If such a switch occurred but the individual did not survey in either house (we've never seen this but it's possible), then if that SID is to be included for some reason (normally one would not), then an ad hoc decision must be made in the data prep script.

Value

Returns a single tbl (tbl_df) with one row for any of the SIDs found in pInLst, and 1 column for each wave, containing the HID the the individual was considered to have been living in for that wave.

[Package RHNetTools version 2.1.3 Index]