makeSurveyedRateTblByHouseByWave {RHNetTools}	R Documentation

Make table showing rates of participation (participants / eligible) x house & wave

Description

Make table showing rates of participation (participants / eligible) x house & wave

Usage


makeSurveyedRateTblByHouseByWave(
  longTBElig,
  longTBSurveyed,
  print2Cons = FALSE
)

Arguments

longTBElig
(tbl_df) A tibble in the form created by 'makeLongNet'; for correct results, this table should include all individuals considered survey-eligible for each wave, whether they took the survey or not. IMPORTANT: The function assumes that this includes individuals showing up only as alters!

longTBSurveyed
(tbl_df) a tibble in the form created by 'makeLongNet'. This input should normally include only individuals who show up as egos in the data. IMPORTANT: the function assumes that alters-only in this table should NOT be counted as 'participants'.

print2Cons
(logical) If TRUE, the output is printed to the console. (Note that this parameter prints only the whole rate table, not the individual elig and surveyed tables)

Value

(tbl_df) A tibble with 5 columns: HID, WID, Elig, Surveyed, and 'Rate' (ordered by WID within HID). 'Elig' shows the number of individuals per house per wave as counted by the union of SID and AltID in 'longTBElig', and normally interpreted to mean, all survey-eligible individuals per house per wave. 'Surveyed' shows the number of individuals per wave as counted by SIDs only, in 'longTBSurveyed'.

Note

It is possible to use this function to calculate rates of anything by house by wave, if you want to, simply by how you filter the two input tables, bearing in mind how this function treats SIDs and AltIDs for each input.

[Package RHNetTools version 2.1.3 Index]