multiGroup {RHNetTools}	R Documentation

Given a list of ID vectors and other arguments accepted by datToSiena(), this function returns a sienaGroup object with groups defined by the list of ID vectors supplied as the first argument.

Description

Given a list of ID vectors and other arguments accepted by datToSiena(), this function returns a sienaGroup object with groups defined by the list of ID vectors supplied as the first argument.

Usage


multiGroup(idlist = NA, ...)

Author(s)

Nathan J. Doogan doogan.1@osu.edu

[Package RHNetTools version 2.1.3 Index]