nIsolates {RHNetTools}	R Documentation

output count of # of isolates in a network

Description

output count of # of isolates in a network

Usage


nIsolates(pNet)

Arguments

pNet
('network') A 'network' class object (from pkg 'network'). Could be binary or non-binary.

Details

Wraps the sna function 'isolates', which gives a list of them, not a count.

Value

(numeric) An integer count of the number of isolates (individuals with no indegrees or outdegrees).

[Package RHNetTools version 2.1.3 Index]