netDiam {RHNetTools}	R Documentation

Find the longest shortest path between any 2 nodes in a network

Description

Find the longest shortest path between any 2 nodes in a network

Usage


netDiam(pNet)

Arguments

pNet
('network') A 'network' class object (from pkg 'network'). Could be binary or non-binary.

Details

Wraps the sna function 'geodist', which gives a matrix of the shortest (directed) path between each pair of nodes.

Value

(numeric) A numeric variable of the max value of this network, ignoring all nodes that are not connected at all.

[Package RHNetTools version 2.1.3 Index]