netVtxAttr {RHNetTools}	R Documentation

Creates a 'network' node attribute

Description

Creates a 'network' node attribute

Usage


netVtxAttr(pAttrMasterDF, pSIDVec)

Arguments

pAttrMasterDF
A two-column data frame or tbl_df with SID and the values of some variable 'x' for (at least) all the SIDs in the desired set, for which there are data (if no data, attribute will be NA). 'x' can be any name you like.

pSIDVec
A vector of SIDs exactly matching the nodes in the desired set for the network of interest. This will be the length of the vertex set (#rows in the returned DF or tbl_df), i.e. '<networkObj>$oel'.

Value

A n x 2 data frame or tbl_df (whichever was passed in), where n is the number of nodes (vertices) to be assigned the value of variable x, col 1 is SID, and col 2 is x.

Note

The calling script should check that the number of rows n is equal to '<networkObj>$oel, and that the same ordering is used in both.
Note also that this is a pretty simple function; it really only selects a subset of some variable and makes sure they're ordered by SID.

[Package RHNetTools version 2.1.3 Index]