recodeRlpVar {RHNetTools}	R Documentation

Recodes a vector representing a relationship variable to 0-1

Description

Recodes a vector representing a relationship variable to 0-1

Usage


recodeRlpVar(pInVec, pTHold)

Arguments

pInVec
A numeric vector representing some relationship variable

pTHold
A number from 1 to 5, or -1 to -5.
If positive, the number means 'this value or higher = 1; else 0'.
If negative, the number means 'this value or lower = 0; else 0'.
Default is 0, which will not do any recoding.

Value

Returns a numeric vector of 0's and 1's representing the recoded input vector.

[Package RHNetTools version 2.1.3 Index]