runConChecks {RHNetTools}	R Documentation

Runs consistency checks on network data

Description

Runs consistency checks on network data

Usage


runConChecks(rlpNet, pWave = 0)

Arguments

rlpNet
(tbl_df) A tbl in 'longNet' or 'rlpNet' format. Can include multiple waves but only one will be processed (see 'pWave' doc).

pWave
(scaler, default=0) The wave to select. If default, function assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise the function will throw an error.

Value

A list of tbl_df's containing specifics needed to find and fix any inconsistencies.
Element 1: The SID and house ID of individuals found to be in multiple houses, along with two 0-1 variables 'ego' and 'alt', which are 1 if ego has chosen alters in the house, and/or received choices in the house, respectively.
Element 2: The House ID, SID, and AltID of any rows in tbl_df that are duplicated
Element 3: Rows of the input tbl_df that are self-selections (loops)
In addition, console output is generated, summarizing the results

[Package RHNetTools version 2.1.3 Index]