s0Fill {RHNetTools}	R Documentation

Finds SIDs of anyone in 2 or more different houses

Description

Finds SIDs of anyone in 2 or more different houses

Usage


s0Fill(netSet)

Arguments

netSet
A 'networkSet' object, equivalent to the output of the 'makeNetworkSet' function, i.e. a list structured as: 1. w 'matrix' objects of equal dimension n, with SIDs as row and column names; 2. One n x w+1 'tbl', sorted on the first column (SID) which gives + The SID of each individual in the entire network set as col 1; + Additional columns HID1, HID2, ...HID<w> giving the house membership ID for each one.

pInTB
A tibble representing an edge-list multiwave set of networks, i.e., with at least variables SID, AltID, HID, WID. Any other vars are ignored. The output of 'makeLongNet' works for this, as does the object generated by 'getNet4OneRlp'

Fills a 'networkSet' of networks with structural zeros between houses

Details

At each wave, the wth adjacency matrix has structural zeros (10) added for all pairs of individuals who are not in the same house for that wave.

Value

A 'networkSet' object identical to the input network Set, but with structural zeros added (as explained in 'details')

[Package RHNetTools version 2.1.3 Index]