selectHouses {RHNetTools}	R Documentation

Select or remove a specified set of houses from a network

Description

Select or remove a specified set of houses from a network

Usage


selectHouses(rlpNet, pHIDVec)

Arguments

rlpNet
(tbl_df) in in the form created by 'makeLongNet', and meant to represent a network edge list. Will actually accept any such input containing at least the variables SID, HID, WID, and AltID. May be single- or multi-wave

pHIDVec
(numerical vecto) Vector of House ID numbers. They must be either all positive numbers, or all negative, or an error is thrown. If all positive, these house IDs are to be included. If all negative, these house IDs are to be excluded

Value

A tbl_df in the same format as the input, including only the user-specified set of houses from pHIDVec

[Package RHNetTools version 2.1.3 Index]